+++ install-stubs/Orchid/Screens/User/UserEditScreen.php
@@ -8,6 +8,7 @@ use App\Orchid\Layouts\User\UserEditLayout;
 use App\Orchid\Layouts\User\UserRoleLayout;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Orchid\Platform\Models\User;
 use Orchid\Screen\Fields\Password;
 use Orchid\Screen\Layouts;
@@ -67,6 +68,7 @@ class UserEditScreen extends Screen
                     Link::name(__('Change Password'))
                         ->icon('icon-lock-open')
                         ->title(__('Change Password'))
+                        ->method('changePassword')
                         ->modal('password'),
                 ]),
 
@@ -92,9 +94,10 @@ class UserEditScreen extends Screen
 
             Layouts::modals([
                 'password' => Layouts::rows([
-                    Password::make('user.password')
+                    Password::make('password')
                         ->title(__('Password'))
-                        ->placeholder('********'),
+                        ->required()
+                        ->placeholder(__('Enter your password')),
                 ]),
             ]),
         ];
@@ -171,4 +174,21 @@ class UserEditScreen extends Screen
 
         return back();
     }
+
+    /**
+     * @param User $user
+     * @param Request $request
+     *
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function changePassword(User $user, Request $request)
+    {
+        $user->password =  Hash::make($request->get('password'));
+        $user->save();
+
+        Alert::info(__('User was saved'));
+
+        return back();
+    }
+
 }